pyproject.toml
[tool.poetry]
name = "Digits"
version = "0.1.0"
description = ""
authors = ["sampath <118274231+sampath017@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.25.2"
pillow = "^10.0.0"
torch = "^2.0.1"
torchvision = "^0.15.2"
torchmetrics = "^1.0.3"
wandb = "^0.15.8"
lightning = "^2.0.7"
extentions = {git = "https://github.com/sampath017/extentions.git"}
gradio = "^4.14.0"
seaborn = "^0.13.1"


[tool.poetry.group.dev.dependencies]
matplotlib = "^3.7.2"
jupyter = "^1.0.0"
notebook = "^7.0.2"
autopep8 = "^2.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
